Make Population Data Type in Listable Long The int datatype was unsuitable because its maximum value of ~2 billion is less than global population

+++ app/src/main/java/org/codethechange/culturemesh/Listable.java
@@ -4,5 +4,5 @@ import android.view.View;
 
 public interface Listable {
     String getListableName();
-    int getNumUsers();
+    long getNumUsers();
 }
